fix(beatport): corregir escritura de BPM en archivos MP3
- Fix Tags ahora usa id3 crate para escribir frame TBPM en MP3
- lofty tiene bug donde escribe BPM pero no puede leerlo después
- Agregada función write_bpm_mp3() en beatport/tagger/writer.rs
- MetadataExtractor mantiene fallback dual (lofty -> id3) para lectura
- Tipo BPM cambiado a Option<f64> para decimales (128.5, 174.23)
- Agregados tests de verificación (test_bpm.rs, test_write_bpm.rs)
- 166 tests backend pasando, 618 tests frontend pasando

BREAKING CHANGE: BPM ahora es Option<f64> en lugar de Option<i32>

Fixes #BPM-not-saving

+++ src-tauri/src/library/metadata/models.rs
@@ -24,7 +24,7 @@ pub struct TrackMetadata {
     pub genre: Option<String>,
 
     /// BPM (beats por minuto) - extraído de tag TBPM/tmpo
-    pub bpm: Option<i32>,
+    pub bpm: Option<f64>,
 
     /// Tonalidad musical (Initial Key) - extraído de tag TKEY/key
     pub key: Option<String>,
